Search Tool
Made the controller to search books in the search bar. Any user can search by title and he will be directed to the particular book page, and then he can look for the author genre or even a partial part of the title, author or genre, and he will have a list that match the search result.