Add and Remove book to favorites
This commit change the particular page of a book, adding next to the title a photo, and under it the button to add or remove the book from favorites depending on if the user is logged and if the book is already a fav book or not. Both buttons are differents.